back_end/models/user_profile/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.1
pydantic==1.10.12
python-jose==3.3.0
pytest
